Add files via upload
Fixed the issue of the robot moving a bit wobbly along the line, it now follows the line much more smoothly.
I achieved this by changing the amount the robot turns each frame as it was overcorrecting quite a bit, I also extended the area that the robot finds white pixels in to be the bottom 10th of the frame instead of one line. I also just overall tidied up the code and made a couple things more efficient.
I also added full commenting to the code.