Updated Core/Completion
This code is an updated version of the previous code uploaded although that version did work on core, it didn't work very well so this is a more purposely designed code for completion. I have changed the program so that the robot's view is split into thirds so that in the middle third it will follow the line normally as it did in the core code but if there is white detected in the left third it will move the robot in that direction along with the values from the center third as well and the same for the right direction. Turning left will have priority over anything else and then straight, then right so the robot will never get stuck or end up going backward. And if there is no white detected and it is a dead-end it will keep turning around until it finds a line to follow.
I have also fixed a few bugs that were happening because of the way that we were getting and processing the frames so the robot should follow the line smoothly now without any errors.